% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of fatalities in a State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number as coded in the \code{maps} package.}

\item{year}{A scalar year.}
}
\value{
This function draws a map of the State and places points on it --
        one point for EACH fatality that occurred in that State.
}
\description{
This function can be used to create a map of fatalities in a US State
   during a year as reported in the FARS database.
}
\details{
This function takes the state number and a year as inputs and
         performs the following operations on them:
         - calls the supporting function, \code{make_filename()}, with
           the year to generate a list of MONTH and year,
         - calls the supporting function, \code{fars_read()}, to read the
           FARS data file with the generated filename,
         - If the state number (provided as the input parameter) is NOT
           found in the FARS data file read: stops the execution with a
           message, "invalid STATE number: xxx".
         - For a valid STATE number:
         - calls the \code{filter()} function from the \code{dplyr}
           package to extract the data for the STATE number provided,
         - If no data exists for the state number: prints a message,
           "no accidents to plot" and returns NULL.
         - Otherwise:
         - sets the invalid data for longitude (values greater than 900)
           and latitude(values greater than 90), if any, to NA, and
         - with the cleaned, mappable data:
         - calls the \code{map()} function of the \code{maps} package to
           draw the map of the State, and
         - calls the \code{points()} function of the base \code{graphics}
           package to draw points on the longitude, latitude coordinates
           as documented in the FARS dataset for that year in that State.
}
\note{
Following functions need to be imported:
      \code{dplyr} package: \code{filter()}.
      \code{maps} package: \code{map()}.
      \code{graphics} base package: \code{points()}.
      Some validaiton is performed in this function as described in the
      "Details" section.
}
\examples{
fars_map_state(1, 2014) # Fatalities map for State #1 during 2014

}

