% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a CSV (Comma Separated Value) file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of the csv file to be
read, residing in the current working directory. CSV file can also be
compressed as allowed by the \code{readr} package's \code{read_csv()}
function.}
}
\value{
This function returns an object of class "tbl_df" (table dataframe)
         from the \code{dplyr} package.
}
\description{
This functions reads a flat file (text file encoded in ASCII) in the current
     working directory and returns it as an object of the "tbl_df" class. If
     the file does not exist in the current directory, this function stops
     execution with an error message. The progress while the file is being
     read is not displayed. Similarly, messages are also suppressed.
     It is an internal (supporting) function used by other functions in this
     package.
}
\details{
This function uses the \code{read_csv()} function from the
         \code{readr} package, with \code{progress = FALSE} parameter.
         After reading the data from the csv file, the \code{tbl_df()}
         function from the \code{dplyr} package is called to convert the
         data into the tabl_df format before returing it.
         This function is also exported, so users will have direct access to
         it when they load the package conataining this function.
}
\note{
The file passed as a parameter (\code{filename}) must exist in the
       current working directory. Otherwise, this function stops execution
       with an error message: "file '\code{filename}' does not exist".
}
\examples{
fars_read("accident_2014.csv.bz2") # file in current working directory
fars_read("abc123xyz.csv") # Error: File 'abc123xyz.csv' does not exist.

}

