% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data for Multiple Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
This function returns an R list comprising of vectors (columns) of
        MONTH and year.
}
\description{
This function takes a vector of years as its only parameter, reads the FARS
   files for those years and returns the MONTH and year columns as a list.
   It is an internal (supporting) function used by other functions in this
   package.
}
\details{
This function processes each element of a vector or list of years
         as follows:
         - calls the supporting function, \code{make_filename()}, to
           generate a filename with the year,
         - calls the supporting function, \code{fars_read()}, to read the
           FARS data file with the generated filename,
         - if the file exists: creates a new column, year, selects the
           MONTH and year columns, and adds them to an R list.
         - If the file does not exit: gives a warning message, and
           adds nothing (NULL) to the list.
         - Returns the R list to the calling function.
}
\note{
Following functions need to be imported:
      \code{dplyr} package: \code{select()} and \code{mutate()}
      Catches any execution error in a generic \code{error()} function and
      ignores that invalid year by returing a NULL.
}
\examples{
fars_read_years(c(2014, 2015, 2016))
fars_read_years(2014:2016)
fars_read_years(list(2014, 2015, 2016))

}

