% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data for multiple years into columns}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
This function returns a dataframe comprising of years (provided as
        the input parameter) as columns with the fatalities summarized by
        each month.
}
\description{
This function can be used to get a "tidy" data frame of FARS data for
   multiple years. Data is summarized by each month of each year with
   each year in its own column.
}
\details{
This function processes each element of a vector or list of years
         as follows:
         - calls the supporting function, \code{fars_read_years()}, to
           generate a list of MONTH and year,
         - calls the \code{bind_rows()} function from the \code{dplyr}
           package to merge the list elements (MONTH and year) into a
           single dataframe,
         - calls the \code{group_by()} function from the \code{dplyr}
           package on the dataframe to group them by year and MONTH,
         - calls the \code{summarize()} function from the \code{dplyr}
           package on the grouped data to count the number of occurrences
           of fatalities for each month within each year, and finally
         - calls the \code{spread()} function from the \code{tidyr}
           package on the summarized data to create a column for each year.
         - Returns the tidy dataframe to the calling function.
}
\note{
Following functions need to be imported:
      \code{dplyr} package: \code{bind_rows()}, \code{group_by()}, and
                            \code{summarize()}.
      \code{tidyr} package: \code{spread}.
      No error handling is performed in this function.
}
\examples{
\dontrun{
fars_summarize_years(c(2014, 2015, 2016))
fars_summarize_years(c(2014:2016))
fars_summarize_years(list(2014, 2015, 2016))
}

}

