% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate the filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, or a value coercible to an integer. If not
coercible, then an invalid filename will be generated containing "NA"
as the year in the file name -- accident_NA.csv.bz2 -- which doesn't
exist in the FARS database.}
}
\value{
This function returns a character string representing the filename.
}
\description{
This function generates the filename for a FARS (Fatality Analysis Reporting
    System) datafile for the year passed as the calling parameter.
    It is an internal (supporting) function used by other functions in this
    package.
}
\details{
This function uses the year, input parameter, to generate the file
         name for the FARS data file. It converts the year into an integer
         and then uses the sprintf() function to format the filename by
         prefixing "accident_" to the year and siffixing ".csv.bz2".
}
\note{
No other input validation is performed. It requires no other packages
         to be imported.
}
\examples{
make_filename(2016) # generates "accident_2016.csv.bz2"
make_filename("2016") # generates "accident_2016.csv.bz2"
make_filename(2016.5) # generates "accident_2016.csv.bz2"
make_filename("2016.5") # generates "accident_2016.csv.bz2"
make_filename("DEC-2016") # generates "accident_NA.csv.bz2"

}

